Merged PR 69775: AVS framework - Initial PR, WIP
## Why is this change being made?
Adding AVS drive/client framework.  Initial development, so a WIP.
## What is changed?
Adding module.
## How is the change being tested?
This was tested and will run on the SVP when avs_isr_test() and associated calls are uncommented.  But this proof of concept was local and didn't invoke the AVS silibs module.

Unit tests will be added.
- [ ] No test, document update only
- [X] Unit tests pass
- [ ] Integration tests pass
- [ ] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [X] SVP
- [ ] FPGA (Add FPGA Config)
- [ ] Silicon
- [ ] Others, explain:

## Notes

## Risks

Related work items: #1721838