Merged PR 71290: fix: update accelerator_ip library to fix crash on FPGA
## Why is this change being made?
scp_accelerators_init, when called on FPGA caused the SCP core to crash. This update is intended to fix this. This is mainly due to the fact that the library is tested on SVP, and FPGA requires additional configurations, including PCR setup on both the VAB and SDMSS tower.

## What is changed?
- Updated silibs to promote_kng_silibs_fpga_r13_16
- enable PCR init on both VAB and SDMSS tower
- other minor tweaks to fix init sequence

## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (add test specifics): Run on both SVP and FPGA to ensure accelerators init completes successfully.
- [ ] Others, explain:

### Test Platform

- [x] SVP: Refer SVP logs in the HeartBeat pipleline on this PR.
- [x] FPGA (R13): [FPGA_Test_Log_R13.txt](https://azurecsi.visualstudio.com/2610da6a-0168-48f8-b34a-f1c882cb6660/_apis/git/repositories/8fc5d0fc-7430-45b1-8963-5d075fba03da/pullRequests/71290/attachments/FPGA_Test_Log_R13.txt)
- [ ] Silicon
- [ ] Others, explain:

## Notes

### Residual risks:
Currently, we use mesh init from CMM scripts, bypassing the mesh_init call in scp prod FW. If mesh_init is called from scp prod fw and then scp_accelerators_init is called, any access to the VAB PCR causes a MemManage fault.

Related work items: #1688140, #1760235