Merged PR 74234: fix issue with svp not launching on latest powershell version
## Why is this change being made?

SVP launch parameters are read from a file.  The powershell snippet that does that creates a list.  It was using split.  It appears internally that split behavior was modified.  Prior it split on multiple spaces. With 7.4.x it splits on each space, so it was passing extra null parameters to run_fixed_vdk.py

## What is changed?
Modified the snippet with syntax that works on 5.1, 7.2.19 and 7.4.2

I also added a try catch block around launching and some debug code to print out the simulator output when needed for debug.  That debug code is disabled by default.

## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [x ] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [x ] SVP
- [ ] FPGA (Add FPGA Config)
- [ ] Silicon
- [ ] Others, explain:

Verified the parameter list is the same for all 3 versions, and then launch SVP from all 3 versions.

## Notes

## Risks

Related work items: #1778090