Merged PR 72839: Implement a common VAB initialization library
## Why is this change being made?

All root port subsystems are connected to the mesh through a VAB
(virtualization and aggregation block).

A common library that powers-up and initializes all VABs can be
implemented as the programming sequence for all VABs is exactly the
same.

## What is changed?

Implement a common VAB initialization library that calls into silibs
APIs for VAB initialization.

## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [x] SVP
- [x] FPGA (Add FPGA Config)
- [ ] Silicon
- [ ] Others, explain:

## Notes

## Risks

Related work items: #1513921