Merged PR 76445: init: add support to bring AP Core0 out of reset
## Why is this change being made?
AP Core0 needs to be brought out of RESET by SCP. On FPGA platforms, this is achieved by using the `SCPPowerOnAP.cmm` script. However, on SVP platform, SCP has to do this activity. Else AP Core0 will not boot.

## What is changed?
Added code to bring AP Core0 out of RESET.

## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (add test specifics): On SVP v9
- [ ] Others, explain:

### Test Platform

- [x] SVP: v9
- [ ] FPGA (Add FPGA Config)
- [ ] Silicon
- [ ] Others, explain:

## Notes
This is a stop-gap arrangement until we get the formal code to enable the AP Cores Power-ON sequence and the below line items need this:
- On SVP, we need to test SCP ProdFW image along with AP Baremetal image to test the end-to-end flow from AP to SDMSS with respect to command/descriptor submission.
- We need to enable the CI/CD pipeline which needs this end-to-end flow.

Because this is a stop-gap arrangement, until we get a formal feature checked-in, we are not writing unit-test and/or functional tests for this PR.

Related work items: #1835614