Merged PR 78786: Add Sensor Fifo config, initial cli cmds, and unit tests
## Why is this change being made?
incremental feature udpate

## What is changed?
the hardware fifo's are not enabled by this PR.

This adds the code required to init the data for the fifo's.
3 cli cmds where added to verify basic functionality.
Unit tests were added for the driver and the service.  To be expanded in follow on PR's.

Updated launch.json so select and run unit tests.

## How is the change being tested?

- [ ] No test, document update only
- [ x] Unit tests pass
- [ ] Integration tests pass
- [ x] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [ x] SVP
- [ ] FPGA (Add FPGA Config)
- [ ] Silicon
- [ ] Others, explain:

Used CLI

## Notes

## Risks

Related work items: #1485735