Merged PR 79775: add local unit test command
## Why is this change being made?

The unittest command was updated to process files for system code coverage. It is integrated with the pipeline.  The update created a more accurate view of total coverage.

However, when developing a unit test suite, and reviewing code coverage for that suite only, there is a large delay due to the system coverage processing. It reduces dev efficiencly.

## What is changed?
added a new command localunittest that operates the same as unittest prior to the system code coverage addition.
Turn around time for development is much faster.  Since it is a new command, there is no impact to the system code coverage.

## How is the change being tested?

- [ ] No test, document update only
- [ x] Unit tests pass
- [ ] Integration tests pass
- [ ] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [ ] SVP
- [ ] FPGA (Add FPGA Config)
- [ ] Silicon
- [ ] Others, explain:

## Notes

## Risks

Related work items: #1485734