Merged PR 79637: accel lib: cdedss: SVP WA: SS-Tower init
## Why is this change being made?
HSP ProdFW is not available to configure the CDEDSS Tower.

FPGA provides cmm script to configure this CDEDSS Tower.

However, for SVP, there is no provision to initialize the CDEDSS Tower.

Hence added a Work-around (WA) code in SCP ProdFW to do this to enable the CDED RCiEP to AP core.

## What is changed?
Added WA code in accel lib to initialize the CDEDSS Tower.

## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (add test specifics):
  - Confirmed SVP Simulation output logs which are accessing the BootCFG registers from accel lib.
  - Ran AP Baremetal and verified the D0 CDED RCiEP's Vendor ID, Device ID and Revision ID configured from SCP - accel lib.
- [ ] Others, explain:

### Test Platform

- [x] SVP: v9
- [ ] FPGA (Add FPGA Config)
- [ ] Silicon
- [ ] Others, explain:

Related work items: #1889565