Merged PR 82355: feat: program spinloop at reset into SDM and CDED ITCMs
## Why is this change being made?
MSCP firmware releases the SDM and CDED accelerators from reset without programming the ITCM. In SVP, the TCMs are pre-loaded, so no issues. On FPGA, this is a problem because elfs are only loaded after SCP is booted up: this causes SDM and CDED to run wildly as soon as reset is released.

There is a current workaround to program a spinloop to SDM and CDED TCM in cmm scripts. This functionality is now instead moved over to firmware, so that we can gradually phase out requirements for cmm script to boot Kingsgate FPGA.

## What is changed?
Used silibs accelip_ss_init function to also program a spinloop at reset for SDM and CDED before releasing their resets.

## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (add test specifics): Tested on R19 FPGA, enabling accelerators.
- [ ] Others, explain:

### Test Platform

- [ ] SVP
- [x] FPGA (R19)
- [ ] Silicon
- [ ] Others, explain:

## Notes

## Risks

----
#### AI description  (iteration 1)
#### PR Classification
New feature: Adds a reset spin loop to SDM and CDED ITCMs for FPGA.

#### PR Summary
This pull request introduces a reset spin loop for FPGA platforms and updates package dependencies.
- `src/libs/accelerator_ip/accelerator_ip.c`: Added a reset spin loop and updated initialization parameters for FPGA.
- `tools/packages.arm-eabi-aarch.xml`: Added a new package dependency `kingsgate.fw.scripts` version `0.0.1`.
- `tools/packages.xml`: Removed the package dependency `kingsgate.fw.scripts` version `0.0.0`.

Related work items: #1947938