Merged PR 83757: fix: SDM emCPU bootup on FPGA R21
## Why is this change being made? What is changed?
SDM emCPU failed to boot on FPGA R21 with the latest main (both sideloading and flash boot failed). This was caused by:
1. An additional parameter `release_m7_halt` required by silibs update to `accelip_ss_init`.
2. Updated spinloop to be written to SDM ITCM.

Both of these are fixed in this PR.

## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (add test specifics): Test SDM emCPU bootup on FPGA R21.
- [ ] Others, explain:

### Test Platform

- [ ] SVP
- [x] FPGA (R21)
- [ ] Silicon
- [ ] Others, explain:

## Notes

## Risks

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix for SDM emCPU bootup on FPGA R21.

#### PR Summary
This pull request addresses the bootup issue of the SDM emCPU on FPGA R21 by updating the spin loop and initialization configurations.
- `src/libs/accelerator_ip/accelerator_ip_configs.c`: Updated `sdm_cded_spin_loop` array to include additional instructions and modified `sdm_emcpu_init_cfg` to set `release_m7_halt` to true for both SDMSS and CDEDSS contexts.

Related work items: #1895361