Merged PR 86465: Add Telemetry In-band data definitions and schema
## Why is this change being made?

Add definitions for the telemetry structures for in-band telemetry reports as well as the schema that is exported for the decoder.

Note: As the Kingsgate spec is not complete, this PR just ports the Pioneer definitions which are expected to remain unchanged.  New data will be added at a later date

## What is changed?

## How is the change being tested?

- [ ] No test, document update only
- [x] Unit tests pass
- [ ] Integration tests pass
- [ ] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [ ] SVP
- [ ] FPGA (Add FPGA Config)
- [ ] Silicon
- [ ] Others, explain:

## Notes

## Risks

----
#### AI description  (iteration 1)
#### PR Classification
New feature: Add telemetry in-band data definitions and schema.

#### PR Summary
This pull request introduces new data structures and schema definitions for in-band telemetry, focusing on power and performance telemetry events.
- `src/services/telemetry/in_band_tlm_cmpnt/inc/telemetry_package_defs.h`: Added definitions for telemetry events, including power telemetry events (e.g., core P-state, core C-state, core throttle) and performance telemetry events (e.g., core, SOC rails, DIMM runtime).

Related work items: #1751050