Merged PR 92573: Add telemetry service DDR management
## Why is this change being made?

telemetry service sends data to the host via DDR

## What is changed?

Primary change is that block pools were created in DDR for power telemetry packages and instantaneous packages.  Power package was enabled and every report time a block is allocated.  Data collection service is not yet implemented, so the next package time, the previous block is freed. This is temporary behavior until DCS is implemented. Verified this operation.

Event traces were added for telemetry service.

Some renaming was performed to align with recent tlm spec naming changes.

## How is the change being tested?

- [ ] No test, document update only
- [x] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (add test specifics)
- [ ] Others, explain:

Verified on both SVP and big FPGA that the report block is allocated and written to.  It is then freed when the next package write occurrs.

### Test Platform

- [x] SVP
- [x] FPGA (Add FPGA Config)
- [ ] Silicon
- [ ] Others, explain:

## Notes

## Risks

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request introduces DDR package buffer management for the telemetry service, addressing the need to allocate and de-allocate package buffers and manage them effectively.
- `src/services/telemetry/in_band_tlm_cmpnt/package_creation.c`: Added functions to enable/disable power and instantaneous telemetry records, and refactored package creation functions to handle new buffer management logic.

Related work items: #2025876