Merged PR 90954: Units : added units for power telemetry elements
## Why is this change being made?
This PR adds the units to different power telemetry elements  used in power telemetry service.   We need to have the units defined for all the power telemetry elements to make have a better understanding to calculate the aggregate values at different layer in the telemetry service as well have a better code readability.
## What is changed?
changed variables names with units extension for telemetry elements and code comments
## How is the change being tested?

- [X] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [ ] Manual tests pass (add test specifics)
- [X] Others, explain:
  - code tested with general boot on SVP.

### Test Platform

- [X] SVP
- [ ] FPGA (Add FPGA Config)
- [ ] Silicon
- [X ] Others, explain:
This PR only add the units to different variables used in repo and code comments with explanations.

## Notes

## Risks

----
#### AI description  (iteration 1)
#### PR Classification
Code cleanup to add units to telemetry items for power metrics.

#### PR Summary
This pull request standardizes the units for various telemetry items related to power metrics.
- `telemetry_package_defs.h`: Added units to fields such as `cstate_entry_latency`, `cstate_exit_latency`, `throttle_priority_residency`, and various frequency fields.
- `tlm_logger.c`: Minor comment updates to reflect changes in telemetry data handling.

Related work items: #2023379