Merged PR 99664: Update SVP Gui Launch command
## Why is this change being made?

In debugging with the SVP team, it turns out that the way we are launching Virtualizer Studio for gui debugging  is incorrect.  It should not be using the run_fixed_vdk.py script for Gui launch.
Removed -UseGui option from runsvp.   runsvp headless operation works the same as before.
Added new alias,  runsvpgui

The gui launch flow was created in its own file to prevent disruption to headless operation.  The gui launch script snippets are borrowed from Braga.

SVP launch is somewhat faster as the vpcfg files are only generated when they need to be.
Restart VpConfig is now reliable.  Prior to this change, the sim would crash for many which increased turnaround time to spin up the sim again.

## What is changed?

described above

## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [x Manual tests pass (add test specifics)
- [ ] Others, explain:

Manual testing with multiple launches of SVP with different configs.

### Test Platform

- [x] SVP
- [ ] FPGA (Add FPGA Config)
- [ ] Silicon
- [ ] Others, explain:

## Notes

## Risks

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request updates the SVP GUI launch command to correct the way Virtualizer Studio is launched for GUI debugging.
- Added new file `tools/pwsh/modules/SVP-GUI-Utils.psm1` with utilities for GUI simulation.
- Removed `-UseGui` option from `Invoke-Virtualizer` function in `tools/pwsh/modules/SVP-Utils.psm1`.
- Added `Get-SvpGuiHelp` function call in `start.ps1` to provide help information for running SVP in GUI mode.

Related work items: #2246432