Merged PR 100525: store nuget temp files local to .externs, cleanup pip cache after pipeline
## Why is this change being made? What is changed?

FPGA Agents have been running out of space due to pip cache buildup on the 1pchie-automation user ID as well as because nuget temporary cache is on C Drive. This PR cleans up pip cache after every pipeline, and configures the nuget temp folder to be within `.externs` so that all temp files are cleaned up automatically after the pipeline.

## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (add test specifics): Tested on FPGA agent that pip cache is purged after the pipeline. Also tested that nuget temp folders are created within the `.externs` directory
- [ ] Others, explain:

### Test Platform

- [ ] SVP
- [x] FPGA
- [ ] Silicon
- [ ] Others, explain:

## Notes

## Risks

----
#### AI description  (iteration 1)
#### PR Classification
Code cleanup

#### PR Summary
This pull request aims to improve the handling of temporary files and cache cleanup in the build pipeline.
- `fpga_pythia_test.yaml`: Added a step to clean up the pip cache after the pipeline runs.
- `Build-Utils.psm1`: Configured NuGet to store temporary files in a local directory within `.externs` instead of the user's profile directory.

Related work items: #2262969