Merged PR 99891: Set Fnc Pointer for Timer
## Why is this change being made?
Set Fnc Pointer for Timer to use in sleep_us function
## What is changed?
Consume new Cortex_M7 Repo release which has a set function pointer for freq/counter
This allows the sleep_us function in Cortex_M7 repo to consume the freq/counter functions in the GTimer module
Add dependency for i3c_controller to wait till gtimer module is run
## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (test on svp, fpga)
- [ ] Others, explain:

### Test Platform

- [x] SVP
- [x] FPGA (Single Die FPGA)
- [ ] Silicon
- [ ] Others, explain:

## Notes
Note that there is a Silibs update for Gtimer as well for the calculations.
Until the silibs PR is not completed, the FPGA will flag an error when running the Gtimer module for the scaling_factor check (value of 1 is currently asserted)
https://dev.azure.com/ms-tsd/Kingsgate/_git/silibs/pullrequest/236281

## Risks

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request introduces a new feature to set function pointers for the timer.
- `src/libs/gtimer/gtimer.c`: Added function pointer setup for `gtimer_prodfw_get_frequency` and `gtimer_prodfw_get_counter`, and added documentation comments.
- `src/libs/gtimer/inc/gtimer_prodfw.h`: Added documentation comments for `gtimer_prodfw_get_counter`.
- `src/libs/runtime_init/i3c_controller_init.c`: Added `gtimer` as a dependency for the `i3c_controller` initialization.
- `src/libs/gtimer/CMakeLists.txt`: Linked `silibs_platform` library.
- `tools/packages.xml`: Updated `cortex_m7` repository tag.

Related work items: #2251971