Merged PR 104812: Create Combined Manifest in DDR for Event Trace / Telemetry Decoding
## Why is this change being made?
the manifest is read by the host to be used by the diagnostic decoder tool to create decoded telemetry.

The final design requires some HSP support that is not currently in place, so this PR creates the manifest for SCP and MCP from the elf file.  Other PR's will add the manifest to flash and then send commands to HSP to load the manifest data into DDR staging.  So some of this will be written at that time.

## What is changed?
Create a combined SCP/MCP manifest in DDR.  Wire up the DCP command to retrieve it.

## How is the change being tested?

Ran on both SVP and big FPGA.  Inspected the MCP's atu mapped address for the combined manifest.
Manual inspection looked correct.  Also verified AP's 64 bit address of the same location and verified the data was there as well.

A follow on task will update diagnostic decoder, which will provide detailed verification of the manifest.

- [ ] No test, document update only
- [x] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [x] SVP
- [x] FPGA (Add FPGA Config)
- [ ] Silicon
- [ ] Others, explain:

## Notes

## Risks

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request introduces the creation of a combined manifest in DDR for event trace and telemetry decoding.
- Added `src/services/data_collection/src/dcs_manifest.c` to handle metadata preparation for manifests.
- Updated `src/libs/in_band_telemetry/diag_decoder.h` with new structures and constants for manifest handling.
- Added unit tests in `tests/unit/services/data_collection/test_dcs_manifest.cpp` to validate the new manifest creation and handling functions.
- Modified `src/libs/in_band_telemetry/CMakeLists.txt` to conditionally compile the `in_band_telemetry_ddr` library based on the platform.

Related work items: #2145303