Merged PR 105952: Update how build_id is generated
## Why is this change being made?
Previously a common cmake was used to generate a random build id value, but it was the same value for both MCP and SCP.  These need to be unique for telemetry and event decoding.

The linker was already generating a unique value per core.  Crash dump was using that.  I moved those definitions in to the common build_data header so there is only 1 location to maintain.

## What is changed?
new method for unique id.  Linker generates 20 bytes.  Existing libraries only have space for 16, use the first 16 bytes from the linker id in that case.

## How is the change being tested?

Dumped the section from both MCP and SCP elfs.  Verified they are unique.
Inspected DCS headers and verified the first 16 bytes of the linker id are used.

- [ ] No test, document update only
- [x] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [x] SVP
- [ ] FPGA (Add FPGA Config)
- [ ] Silicon
- [ ] Others, explain:

## Notes

## Risks

----
#### AI description  (iteration 1)
#### PR Classification
Code update to change the method of generating `build_id`.

#### PR Summary
This pull request updates the method of generating `build_id` to align with new requirements and improve consistency.
- Modified `src/libs/build_data/build_data.h.in` to define a new structure `NOTE_GNU_BUILD_ID` and removed `BUILD_PRODUCT_GUID`.
- Added `src/libs/build_data/win32/build_data_dummy.c` for unit testing purposes.
- Updated `src/services/data_collection/src/dcs_manifest.c` to use the new `g_note_gnu_build_id` for manifest ID generation.
- Adjusted `src/libs/crash_dump/src/crash_dump_payload_msft.c` to use the new `g_note_gnu_build_id` for build ID.
- Updated linker scripts in `tools/cmakes/toolchain/arm-eabi-aarch/ld/kernel.ld` and other related files to include the new build ID structure.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #2145453